requirements.txt
streamlit
PyPDF2
langchain
langchain_google_genai
google-cloud
google-api-core
google-generativeai
faiss-cpu
python-docx
gtts
python-dotenv